+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - tests directory to run a single inference test
 - cmake rules to install JHU repos
+- goes specific preprocess script
 ### Changed
 
 ### Fixed